:recycle: Migrate review submitted estimate step to newer pattern.
