--- web/src/pages/my-travel-requests/estimate/EstimatePage.vue
@@ -1,46 +1,0 @@
-<template>
-  <div>
-    <EstimatesTable
-      ref="estimatesTable"
-      :travel-authorization-id="travelAuthorizationIdAsNumber"
-    />
-  </div>
-</template>
-
-<script setup>
-import { computed, toRefs } from "vue"
-
-import useBreadcrumbs from "@/use/use-breadcrumbs"
-import useTravelAuthorization from "@/use/use-travel-authorization"
-
-import EstimatesTable from "@/modules/travel-authorizations/components/read-travel-authorization-estimate-page/EstimatesTable"
-
-const props = defineProps({
-  travelAuthorizationId: {
-    type: [String, Number],
-    required: true,
-  },
-})
-
-const travelAuthorizationIdAsNumber = computed(() => parseInt(props.travelAuthorizationId))
-
-const { travelAuthorizationId } = toRefs(props)
-const { travelAuthorization } = useTravelAuthorization(travelAuthorizationId)
-
-const breadcrumbs = computed(() => [
-  {
-    text: "My Travel Requests",
-    to: {
-      name: "my-travel-requests/MyTravelRequestsPage",
-    },
-  },
-  {
-    text: travelAuthorization.value?.eventName || "loading ...",
-    to: {
-      name: "my-travel-requests/estimate/EstimatePage",
-      params: { travelAuthorizationId: travelAuthorizationId.value },
-    },
-  },
-])
-useBreadcrumbs(breadcrumbs)
-</script>
